requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
Click==7.0
Flask==1.1.1
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==7.2.0
mypy==0.740
mypy-extensions==0.4.3
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
SQLAlchemy==1.3.11
typed-ast==1.4.0
typing-extensions==3.7.4.1
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
